backend/requirements.txt
asgiref==3.8.1
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
CodeConvert==3.0.2
cryptography==43.0.0
defusedxml==0.8.0rc2
Django==4.2.16
django-admin==2.0.2
django-cors-headers==4.4.0
psycopg2-binary==2.9.3
django-excel-response2==3.0.6
django-filter==23.1
django-link-shortener==0.5
django-models-ext==1.1.11
django-short-url==1.1.8
django-six==1.0.5
django-templated-mail==1.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
drf-extra-fields == 3.7.0
excel-base==1.0.4
furl==2.1.3
idna==3.7
isoweek==1.3.3
oauthlib==3.2.2
orderedmultidict==1.0.1
pillow==10.4.0
pycparser==2.22
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.2
requests==2.32.3
requests-oauthlib==2.0.0
screen==1.0.1
shortuuid==1.0.13
six==1.16.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
TimeConvert==3.0.13
typing_extensions==4.12.2
tzlocal==5.2
urllib3==2.2.2
xlwt==1.3.0
